requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pypiwin32==223
pywin32==306
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.4
win10toast==0.9
